storage: guard timestamp binding seal frontier from regressing
Storage workers are not smart enough right now to only report new
changes to the timestamp bindings they contain the their respective
frontiers. Consequently we need to guard advancing the stash frontier
that asserts this doesn't happen until we fix the storage workers.

This is not new behavior, we just didn't have the assertion in the
previous way we were storing timestamp bindings, but we should fix it
anyway because once we're sure this never happens the assertion might
catch real bugs.

Fixes #11594

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>