Trigger pod replacement by hash annotation
Kubernetes does not always replace pods when a StatefulSet changes.

Adds an annotation to the PodTemplateSpec containing the sha256 hash of
that PodTemplateSpec. After applying the Statefulset, we get the pod.
If the pod exists but the annotation doesn't match, we delete the pod to
trigger K8S to recreate it.