materialized: allow enabling CORS by origin
In Materialize Platform, the web UI will make requests of `materialized`
directly. It does so from a different origin, so we need to enable
cross-origin resource sharing (CORS) or those requests are blocked by
the browser.

This commit adds a new command line option to `materialized` named
`--cors-allowed-origin` that allows the operator of `materialized` to
specify from what origins CORS requests are allowed.

Touches #11533. This doesn't actually entirely resolve #11533, because
this commit doesn't switch us to using axum; I figured out how to get
the CORS support in with our existing framework.