materialized: swap HTTP server for axum
axum is the Rust HTTP server framework that we've been looking for.
We're using it quite successfully in MaterializeInc/cloud. Use it in
materialized, too. It handles quite a bit of routing and form parsing
that we previously had to do manually.

Forthcoming commits will be able to build on this foundation to provide
a WebSocket API. We'll also be able to use nested routing to split the
HTTP server up into separate crates, so that other binaries besides
`materialized` can provide an HTTP server. For example, `dataflowd`
needs to expose metrics and profiling endpoints.