materialized: add local process orchestrator
Add a new `Orchestrator` backend, `ProcessOrchestrator` that manages
local processes. This sits alongside the existing
`KubernetesOrchestrator` backend.

The `ProcessOrchestrator` is intended for local development. It is much
more convenient to have `materialized` spin up `dataflowd` processes as
necessary than to have to futz with Minikube.

The big defect with this orchestrator is that you have to remember to
build `dataflowd`. If you are not careful, you can end up with
`materialized` and `dataflowd` processes that are out of sync. In other
words, it's a bit too easy to accidentally test version compatibility
while doing local development with this orchestrator. We can fix this
later with e.g. a shell script that builds both dataflowd and
materialized.

To use the process orchestrator, simply run:

    $ cargo build --bin dataflowd
    $ cargo run -- --dev --orchestrator=process

The `materialized` process will automatically launch `dataflowd
--runtime=storage`. You can type

    # CREATE CLUSTER ... SIZE 'ignored'

to spin up separate `dataflowd --runtime=compute` processes.