coord: Add support for FilesystemSecretsController (#11628)
Continuation of the SECRETS work as defined in its design.

This PR adds a configurable Local Filesystem secrets controller. The controller stores the contents of the secret on disk inside of the mzdata/secrets directory. Each secret is stored in a file that matches its GlobalId.

The user can select which secret controller they want to use via --secrets-controller the two options are local-file-system and kubernetes. The default is the FilesystemSecretsController.

This PR also adds the necessary files for the Kubernetes controller, yet the controller does nothing.

Motivation
This PR adds a known-desirable feature.

Co-authored-by: Philip Stoev <pstoev@materialize.io>