[sources] Respect Debezium transaction boundaries (#11306)
Read in Debezium transaction data alongside the data so that we can keep everything from the same transaction at the same timestamp.  We do this with the following way:
- Create a source that contains the transaction data in the format described here: https://debezium.io/documentation/reference/connectors/postgresql.html#postgresql-transaction-metadata
- Create a second source using `WITH tx_metadata = ??`

Internally this second source creates a binary timely operator that reads in both streams of data and matches them up.  On each invocation this operator advances to the `meet` of the tx metadata and data frontiers when there's no outstanding transaction or data to match up.  If either of these are outstanding, we should not advance the frontier.

It may be possible to come up with a tighter bound but I'd like to avoid diving into that too much in this PR -- but please let me know if you think we should tightening it and I'll make a follow-up item

### Motivation
Fixes #8693
Fixes #8695
Fixes #6296
Fixes #6555

### Tips for reviewer
Files by type of changes (hopefully this makes what's going on more clear)
- `src/dataflow/src/render/debezium.rs`: this is where we actually do the merging of the tx metadata and the data
- `src/dataflow/src/render/sources.rs`: this is where we re-create the source that contains the tx metadata
- `src/sql/src/plan/statement/ddl.rs`: this is where we validate the shape of the tx metadata source (including that it has the proper columns) and setting up the dependencies for the data source (e.g. we don't want to allow the metadata source to be dropped if the data source is still alive)
- `test/testdrive/kafka-avro-debezium-transaction.td` new tests!

### Testing

- [x] This PR has adequate test coverage / QA involvement has been duly considered.

### Release notes

This PR includes the following [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):

- `tx_metadata` / `tx_metadata_collection_name` are now accepted as options for debezium sources so that they're able to be transactional.  This lives behind an experimental flag.  Multiple data sources can share a single transaction metadata source