persist: restrict all impls of Consensus to use seqnos in [0, i64::MAX]
This commit teaches Consensus to constrain itself to sequence numbers in the
nonnegative range that can be expressed by signed 64 bit integers as SQL databases
(e.g. Postgres and Sqlite) don't provide 64 bit unsigned natively and its not
worth the complexity to support that exta range.

As part of that, this commit:
- teaches the in-memory and sqlite Consensus implementations to properly error
  if they encounter a sequence number outside of this range.
- adds new unit tests to check the edge cases in this range.