feature-benchmark: use persistent testdrive container
Use docker-compose 'exec' rather than 'run' to run the individual
testdrive snippets that the feature benchmark executes.

This removes a ~0.5s overhead when taking each measurement

As the testdrive container is now persistent across measurements,
TCP port exhaustion occured in the ConnectionLatency scenario. Fix
this by reducing the SCALE for this particular scenario.