storaged: add prof HTTP endpoint too
Add the HTTP profiling server to storaged, too. It was added to computed
in #11882.

This commit pushes more of the routing logic into the prof crate so that
less code is duplicated between storaged and computed. The trick is to
use axum's support for nesting routers.

This commit also tweaks the `build_info` macro so that it is easier to
drop into any crate. It no longer requires that the calling crate depend
on `compile_time_run` explicitly, nor that the calling crate add the
target triple to the environment via a build script.