expr: deprecate `_nolimit` visitor methods
This commit renames the existing infallible visitor methods with a
`_nolimit` suffix and deprecates them. This frees up the old names for
replacement methods that take an infallible callback, but still return a
`Result` to report `RecursionLimitError`s. Going forward, new code
should use the safe methods instead of the old `_nolimit` ones.

This commit does not make attempts to port the existing client code to
the new safe methods and instead adds `#[cfg(allow(deprecation)]`
annotations. The intent is to allow us to migrate the client code over
time, instead of having to spend a lot of effort now.