doc/developer: note that PubNub sources are reclockable
After all this time believing that PubNub sources do not have an
upstream gauge of progress... it actually seems that they do. Each
PubNub message has a "time token" [0], which seems to have exactly the
right properties.

It's not possible to use this time token to seek through a PubNub stream
unless the stream has history enabled, but it should always be possible
to use this token to deduplicate messages. You might still miss
messages, but at least you won't duplicate them.

[0]: https://docs.rs/pubnub-hyper/latest/pubnub_hyper/core/data/timetoken/struct.Timetoken.html