deps: require Python 3.9+ and "dev" dependency classes
Now that we have a separate repository for demos, we no longer need to
support a relatively old version of Python, nor aggressively limit our
dependency set to avoid `pip install` problems on users' machines. See
issue #10054 for details.

This commit tackles the first two items on the todo list:

  * Merging requirements and requirements-dev.txt.

  * Bumping the minimum supported version of Python to v3.9.