materialized: Reconnect to existing processes (#11983)
Previously, when using the process orchestrator, materialized would
always attempt to start new storaged and computed processes. The one
exception is that new computed processes aren't started for remote
clusters. This made it very difficult to restart materialized
independently of computed and storaged processes.

This commit adds a feature to materiazed that attempts to connect to
existing storaged and computed processes. storaged and computed
processes create pid files when they start and delete the files when
they stop. A file containing port mappings is also created and deleted
for each process. materialized will look for these files and connect to
the owning process before starting new processes.

Resolves #11974