repr: remove `custom_name` from `ScalarType::Record`
The `custom_name` field is redundant with the `custom_id` field, so
remove it.

This field was used by Avro sinks to name some well-known record types
in the generated Avro schemas. Pivot this schema generation to use the
custom ID field instead. This requires a bit of a hack to give these
types IDs, because these aren't real types in the catalog yet, but it's
still a net improvement.

Fix #11344.