persist: change WriteHandle::append to return current upper on mismatch
With timeouts, it can happen that a write succeeds but the write handle
doesn't learn about the updated upper. Previously, in those cases, we
would yield an "invalid usage" error. Now, we instead return the current
upper, similar to `compare_and_append()` and also update the `upper` of
the `WriteHandle`.