catalog-compat: Disable the catalog compat test
The upgrade scenarios that were tested by this test are no longer
supported. Furthermore, the 0.7 binary that is being used now
asserts in code that is no longer present in the current Mz version.

The test/catalog-compat directory is being kept at this time so
that any relevant individual test cases from it can be migrated to
a future upgrade framework.