augment storage and compute instance commands with source metadata
In order to decouple storage and compute some amount of metadata must be
communicated to each one through their command streams as a means of
rendezvous.

This is currently just the persist location but in the future it will
include a list of persist shards, the shard for the timestamp bindings
collection, etc.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>