persist: fix correctness bug in snapshot when as_of >= upper
Snapshot was meant to block until the requested data is ready, but
before this fix, if passed an as_of that was >= upper, it would
immediately return whatever it currently had. This is a correctness bug.

While in here, addressed a related TODO about only fetching state if we
can't immediately serve the snapshot.

When I first saw this, I thought a similar issue might be present in
Listen, but I added test coverage and Listen seems to be okay.

Fixes #12131
Touches #11977