sql: add AS OF AT LEAST
AT LEAST functions the same as not specifying AS OF, but additionally
advances to the specified timestamp. Notably, it is not the same as
AS OF but also advancing to the since. We could absolutely do that,
but if we're going to choose a timestamp in advance of what the user
specified, we might as well choose something with the freshness and
table linearizability guarantees we have elsewhere. AT LEAST is a way
for a user to tell the system to ensure it has some known data in it,
but otherwise the caller doesn't care what happens.

Implement this with a refactor of `determine_timestamp`. Policy
is determined by the `QueryWhen` enum, and implementation by
`determine_timestamp`.