doc/developer: update and restructure guide
Update and restructure the developer guide for the Materialize Platform
world. The introduction is reworded to note that Materialize consists of
several services in Rust, plus a bunch of tooling in Python and Bash.
The "Editors and IDEs" and "Debugging" sections are moved under
"Developer tools". The "Code structure" section growsn an overview of
the driectoreis in the repository. The "Publishing code changes" is
moved to a separate document because it is rarely relevant. Various dead
links are removed.