persist: remove timeouts in favor of retrying forever
This commits leans all the way into the rust async idiom of "drop is
cancellation". All timeouts in the persist API are removed in favor of
retrying indefinitely.

A followup commit in this PR will remove nearly all of the ExternalError
Results from the Persist API. Left for future work (after this PR) is
unit test coverage of our drop-cancel propagation (because that's only
an issue around tasks, really).

This commit adds a new invariant to `Location::open` (and thus the
`open` call of various Blob and Consensus impls) to only return once
we've successfully connected. This allows every part of persist after
Location::open to assume that the Blob and Consensus are working (aka
retry indefinitely).

All uses of persist in the codebase have had timeouts removed. We can
add them back in as found necessary.