persist: make `append` functions accept a lot more types
The previous API required that the user came up with an iterator of
references to values but that meant that every called that had owned
values had to write unecessary boilerplate. For example if a user had a
`Vec<((K, V), T, D)>` ready to go they would need to first make this an
iterator over `&((K, V), T, D)` and then use a map adaptor to convert
that to an iterator over `((&K, &V), &T, &D)`.

This patch changes the `append` APIs to accept any combination of owned
and borrowed values (a total of 32 combinations) which makes calling
these APIs much nicer.

The original usecase for this change isn't visible in this PR because it
is part of the larger #12216 PR, but I thought I'd break it up for
easier review. You can see however how the usage becomes nicer in
`persist_open_loop_benchmark.rs` and also in
`src/persist-client/src/lib.rs`.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>