sql: Disable multi-table write transactions (#12290)
Multi-table write transactions introduce new failure modes for
transactions that we don't properly handle. For example append commands
are only atomic over a single table. If append fails after applying the
updates for a subset of the tables in the write transaction, then we
don't have a good way to recover and we can be left in a partially
applied state.

Therefore this commit returns an error when transactions try to write
to multiple tables.

Fixes #12288