persist: make `lower` an explicit argument to `append()`
Before, we were using the (cached) `upper` of the `WriteHandle` as the
implied `lower` of newly appended batches. Now we require the `lower` to
be explicitly specified.

Most call sites are changed to use `handle.upper().clone()` to get the
same behaviour as previously. Some tests do pass in an explicit `lower`
now, though.