scratch: install postgres and set up an ubuntu user
Both stash and persist are backed by postgres when doing local
development, so seems reasonable to have around. Set it up so it's easy
to connect to and export the MZDEV_POSTGRES env var with connection
details.

After this commit, both of the following work on a fresh ./bin/scratch

    psql
    psql "$MZDEV_POSTGRES"