[coord] New secret guarantees (#12334)
We have decided to move the creation of the secret into the environment controller in Kubernetes. When using the Kubernetes environment, all scripts have to create a new user-managed-secrets secret, if they desire to use the Kubernetes controller.

The KubernetesSecretsController now waits until the secret becomes visible on its local filesystem. When altering secrets, we do not guarantee that the new secret content has been propagated.