scratch: update ec2-instance-connect scripts
The ec2-instance-connect package currently available in the Ubuntu
package repo is not compatible with OpenSSL 3, the version installed on
Ubuntu 22.04. The issue is fixed upstream and we are waiting for it to
land in the Ubuntu repo. Until then we perform the update manually.

The proper way to install ec2-instance-connect from source would be
building a deb package and then installing that. This requires a bunch
of build tools and GPG signing keys. Simply copying the scripts to their
install location is easier and should be fine for a temporary
workaround.