tests: Do not rely on the kafka TCP port for unrelated tests
Tests that needed to specify a bogus schema registry URL were using
the kafka TCP port in the expecation that Kafka will quickly close
the socket.

This assumption however is unfortunately not true for Redpanda, where
the socket remains open forever without a discernible timeout, causing
such tests to fail in unexpected ways.

The fix is to remove any tests that were abusing the kafka port in
this manner. There is no other suitable service within the testdrive
composition that such tests could be directed to.