persist: make sqlite consensus impl properly async
It was making blocking io calls inside async fns, which is the single
biggest async no-no. This may be contributing to the slowness that
Petros is seeing in CI on 12216, but either way it needs to be fixed if
SqliteConsensus is going to be used until we can switch to Postgres for
Consensus in dev/CI.

While I'm in here, tweak the string formatting of the queries so that
rustfmt doesn't barf on them quite as much. It's still not great, but I
couldn't come up with anything better than this. Open to suggestions.

On a persist-benchmarking machine

    MZ_PERSIST_RECORD_COUNT_SMALL=64 MZ_PERSIST_RECORD_SIZE_BYTES_SMALL=1024 MZ_PERSIST_BATCH_MAX_COUNT_SMALL=8
    consensus/compare_and_set/sqlite/64KiB
                            time:   [7.8217 ms 7.9570 ms 8.1144 ms]
                            thrpt:  [7.7024 MiB/s 7.8547 MiB/s 7.9906 MiB/s]
                     change:
                            time:   [-3.1472% -0.6147% +2.0225%] (p = 0.64 > 0.05)
                            thrpt:  [-1.9824% +0.6185% +3.2495%]
                            No change in performance detected.