persist: separate writing updates to blob store from appending to shard metadata
An `append()` (or `compare_and_append()`) call internally does two
separate steps:

1. write the batch of updates to a Blob implementation (S3, in
   production)
2. add a pointer to the new batch in the metadata (aka state machine),
   which we keep in a Consensus implementation

We don't currently retry the second step, so a failure can void the
complete append attempt, which has to be retried as a whole.

Now, we allow separating the two parts at the API level to allow writing
the batch of updates separately (which we can already retry internally
in the client) and then retry "appending" it into the shard metadata.

Background context: we don't retry the second step, because we cannot
ensure correctness for some usage patterns.

Closes #12468