materialized: enable PostgreSQL by default for stash
Change bin/materialized and the materialized Docker image to use
PostgreSQL as the default backing for the stash. Because PostgreSQL is
usually configured so that `psql` provides passwordless authentication
to only one database, we use the same database for both persist
consensus and the catalog stash; each gets its own *schema* rather than
database.

The invoker of `materialized` (i.e., the actual Rust binary) is
responsible for creating the PostgreSQL schemas. That means all of the
following learn to create the requisite schemas:

  (1) the `bin/materialized` wrapper script
  (2) the `materialized` Docker entrypoint script
  (3) the MaterializeInc/cloud environment controller (not in this commit)

(1) required rewriting the `bin/materialized` script in Python as
otherwise the required argument parsing got unwieldy.