pgwire: check for conn close while waiting for rows
Long running queries where the client disconnects before the query is
ready would previously keep the query around until it was ready. In the
case of permanently failing queries (like using mz_panic), these would
never return and leak resources. Worse, we would never be able to tell
compute to cancel the query to exit its crash loop (in the mz_panic case
or, say, an OOM loop).

Detect a disconnected client and terminate the connection, which will
propagate a cancel through the rest of the stack.

Fixes #12546