Revert "persist: wrap all pg consensus ops in txns (#12534)"
This reverts commit 60cf8fdebf4088f1e18571e66978d41b8abbdb9b.

The bug that 60cf8fdeb worked around was actually a bug in
tokio-postgres for we'd forgotten to pick up the fix (#8738). Now that
we've picked up the tokio-postgres fix, we can go back to the original
Postgre consensus implementation, which we believe will be a bit faster
in production because it involves fewer network roundtrips.