[coord] Enable multi-statement alter secret transactions (#12592)
It is desirable that a user can alter multiple secrets in one atomic operation. For example, one might want to change the CA along with the server certificate, or client certificate.

The FS SecretsController does not support atomic transactions on multiple secrets, but the Kubernetes controller does.

This PR only allows the mutation of multiple secrets via the alter statement. Neither drop nor create can be mixed in a single transaction.